requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
aiosqlite==0.21.0
alembic==1.16.1
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.9.0
appdirs==1.4.4
arrow==1.3.0
bcrypt==4.3.0
captcha==0.7.1
certifi==2025.6.15
charset-normalizer==3.4.2
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
databricks-sdk==0.55.0
ddgs
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
docker==7.1.0
dydantic==0.0.8
executing==2.2.0
extra-streamlit-components==0.1.80
fastapi==0.115.12
Flask==3.1.1
fonttools==4.58.1
fqdn==1.5.1
google-auth==2.40.2
googleapis-common-protos==1.70.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.3
grpcio==1.73.1
h11==0.16.0
httpcore==1.0.9
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
joblib==1.5.1
jsonpointer==2.1
jupyter-highlight-selected-word==0.2.0
jupyter_client==8.6.3
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.7.2
jupyter_nbextensions_configurator==0.6.4
kiwisolver==1.4.8
langchain-anthropic==0.3.13
langgraph-checkpoint==2.0.26
langgraph-checkpoint-sqlite==2.0.10
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langmem==0.0.26
langfuse==3.0.7
lxml==5.4.0
Mako==1.3.10
Markdown==3.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
mkl-service==2.4.0
mlflow==2.22.0
mlflow-skinny==2.22.0
mypy_extensions==0.4.1
nest-asyncio==1.6.0
networkx==3.5
openai==0.34.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
ormsgpack==1.10.0
parso==0.8.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pyppeteer==2.0.0
PyQt6==6.7.1
python-dateutil==2.9.0.post0
pywin32==310
pyzmq==26.4.0
requests==2.32.4
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
setuptools==72.1.0
six==1.17.0
sniffio==1.3.1
sqlite-vec==0.1.6
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
streamlit-authenticator==0.4.2
test-class==0.0.2
threadpoolctl==3.6.0
tiktoken
toml
tornado==6.5
traitlets==5.14.3
trustcall==0.0.39
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.34.3
waitress==3.0.2
wcwidth==0.2.13
webcolors==24.11.1
websockets==10.4
Werkzeug==3.1.3
wheel==0.45.1
zipp==3.23.0